Fix: keep utility.py and test_utility.py (undo utility→utils merge from cleanup #2285)
The original commit f9ee3e85 accidentally re-merged utility.py into utils.py,
undoing the revert from PR #2324. This fixup restores utility.py and
test_utility.py, reverts the content merge in utils.py/test_utils.py,
and fixes imports back to pymc_marketing.mmm.utility.
